Move getBrazeMetaFromQueryString out of BrazeBanner into shared file
